// frontend/src/app/models/timeTable.ts
export interface TimeTable {
  id: number;
  day: string;
  sectionId: number;
  sectionName?: string;
  yearId: number;
  yearName?: string;
  period1SubjectId: number;
  period1SubjectName?: string;
  period2SubjectId: number;
  period2SubjectName?: string;
  period3SubjectId: number;
  period3SubjectName?: string;
  period4SubjectId: number;
  period4SubjectName?: string;
  period5SubjectId: number;
  period5SubjectName?: string;
  period6SubjectId: number;
  period6SubjectName?: string;
  period7SubjectId: number;
  period7SubjectName?: string;
  period8SubjectId: number;
  period8SubjectName?: string;
}